--- df_objects/__init__.py
@@ -1,1 +1,0 @@
-from df_objects.df_objects import DemandDf, ProductionDf, ElectricityUseDf, CostElectricityDf, SimulationResults
