+++ hourly_simulation/strategies/selling_strategy.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import numpy as np
 import pandas as pd
 
-from df_objects.df_objects import DemandDf, ProductionDf, ElectricityUseDf, CostElectricityDf
+from objects.df import DemandDf, ProductionDf, ElectricityUseDf, CostElectricityDf
 from hourly_simulation.parameters import Params, ELECTRICITY_COST, BINARY_SELLING_COST, ELECTRICITY_SELLING_INCOME
 from hourly_simulation.shift_day_in_year import shift_day_of_year
